fix: Modify code layout and annotations
