+++ app/comment/service/internal/biz/comment.go
@@ -1,6 +1,7 @@
 package biz
 
 import (
+	"Atreus/app/comment/service/internal/conf"
 	"context"
 	"errors"
 	"github.com/go-kratos/kratos/v2/log"
@@ -26,6 +27,23 @@ type User struct {
 	TotalFavorited  uint32
 	WorkCount       uint32
 	FavoriteCount   uint32
+}
+
+type CommentRepo interface {
+	CreateComment(context.Context, uint32, string, uint32) (*Comment, error)
+	DeleteComment(context.Context, uint32, uint32, uint32) (*Comment, error)
+	GetCommentList(context.Context, uint32) ([]*Comment, error)
+	GetCommentNumber(context.Context, uint32) (int64, error)
+}
+
+type CommentUsecase struct {
+	commentRepo CommentRepo
+	config      *conf.JWT
+	log         *log.Helper
+}
+
+func NewCommentUsecase(conf *conf.JWT, cr CommentRepo, logger log.Logger) *CommentUsecase {
+	return &CommentUsecase{config: conf, commentRepo: cr, log: log.NewHelper(log.With(logger, "model", "usecase/comment"))}
 }
 
 // parseToken 接收TokenString进行校验
