+++ app/comment/service/internal/conf/conf.proto
@@ -41,11 +41,11 @@ message Data {
     string dsn = 2;
   }
    message Redis {
-      int32 db = 1;
-     string addr = 2;
-     string password = 3;
-     google.protobuf.Duration read_timeout = 4;
-     google.protobuf.Duration write_timeout = 5;
+    int32 db = 1;
+    string addr = 2;
+    string password = 3;
+    google.protobuf.Duration read_timeout = 4;
+    google.protobuf.Duration write_timeout = 5;
    }
   Mysql mysql = 1;
   Redis redis = 2;
