--- deploy/docker-compose.yaml
@@ -1,37 +1,0 @@
-version: '3'
-services:
-  mysql:
-    image: mysql:8.0.34
-    restart: always
-    container_name: atreus_db
-    environment:
-      MYSQL_ROOT_PASSWORD: "toomanysource"
-      MYSQL_DATABASE: "atreus"
-      MYSQL_USER: "atreus" #自定义数据库的用户，权限只作用于MYSQL_DATABASE配置的数据库
-      MYSQL_PASSWORD: "atreus"
-      TZ: "Asia/Shanghai"
-    ports:
-      - "3306:3306"
-    volumes:
-      - ../data/mysql/data:/var/lib/mysql
-      - ../data/mysql/conf/my.cnf:/etc/my.cnf
-      # 数据库还原目录 可将需要还原的sql文件放在这里
-      - ../data/mysql/source:/docker-entrypoint-initdb.d
-  redis:
-    image: redis:7.0.12
-    container_name: atreus_cache
-    environment:
-      TZ: "Asia/Shanghai"
-    volumes:
-      - ../data/redis/data:/data
-      - ../data/redis/conf/redis.conf:/usr/local/etc/redis/redis.conf
-      - ../data/redis/logs:/logs
-    command: redis-server --requirepass atreus
-    #   #  两个写入操作 只是为了解决启动后警告 可以去掉
-    #    /bin/bash -c "echo 511 > /proc/sys/net/core/somaxconn
-    #    && echo never > /sys/kernel/mm/transparent_hugepage/enabled
-    #    && redis-server /usr/local/etc/redis/redis.conf"
-    ports:
-      # 使用宿主机的端口映射到容器的端口
-      # 宿主机：容器
-      - 6379:6379
